+++ noodles-vcf/src/record/genotypes.rs
@@ -6,8 +6,9 @@ pub mod keys;
 pub use self::{genotype::Genotype, keys::Keys};
 
 use std::{
-    fmt,
+    error, fmt,
     ops::{Deref, DerefMut},
+    str::FromStr,
 };
 
 use self::genotype::field;
@@ -75,6 +76,56 @@ impl fmt::Display for Genotypes {
     }
 }
 
+/// An error returned when raw VCF record genotypes fail to parse.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub enum ParseError {
+    /// The input is empty.
+    Empty,
+    /// The input is invalid.
+    Invalid,
+    /// A key is invalid.
+    InvalidKeys(keys::ParseError),
+    /// A genotype is invalid.
+    InvalidGenotype(genotype::ParseError),
+}
+
+impl error::Error for ParseError {}
+
+impl fmt::Display for ParseError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Empty => f.write_str("empty input"),
+            Self::Invalid => f.write_str("invalid input"),
+            Self::InvalidKeys(e) => write!(f, "invalid keys: {}", e),
+            Self::InvalidGenotype(e) => write!(f, "invalid genotype: {}", e),
+        }
+    }
+}
+
+impl FromStr for Genotypes {
+    type Err = ParseError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        use super::FIELD_DELIMITER;
+
+        if s.is_empty() {
+            return Err(ParseError::Empty);
+        }
+
+        let (format, t) = s.split_once(FIELD_DELIMITER).ok_or(ParseError::Invalid)?;
+
+        let keys = format.parse().map_err(ParseError::InvalidKeys)?;
+
+        let genotypes = t
+            .split(FIELD_DELIMITER)
+            .map(|t| Genotype::from_str_format(t, &keys))
+            .collect::<Result<_, _>>()
+            .map_err(ParseError::InvalidGenotype)?;
+
+        Ok(Self::new(keys, genotypes))
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -119,6 +170,36 @@ mod tests {
         );
 
         assert_eq!(genotypes.to_string(), "GT:GQ\t0|0:13");
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_from_str() -> Result<(), Box<dyn std::error::Error>> {
+        use genotype::{
+            field::{Key, Value},
+            Field,
+        };
+
+        let expected = Genotypes::new(
+            Keys::try_from(vec![Key::Genotype, Key::ConditionalGenotypeQuality])?,
+            vec![Genotype::try_from(vec![
+                Field::new(Key::Genotype, Some(Value::String(String::from("0|0")))),
+                Field::new(Key::ConditionalGenotypeQuality, Some(Value::Integer(13))),
+            ])?],
+        );
+        assert_eq!("GT:GQ\t0|0:13".parse(), Ok(expected));
+
+        assert_eq!("".parse::<Genotypes>(), Err(ParseError::Empty));
+        assert_eq!("GT:GQ".parse::<Genotypes>(), Err(ParseError::Invalid));
+        assert!(matches!(
+            "\t".parse::<Genotypes>(),
+            Err(ParseError::InvalidKeys(_))
+        ));
+        assert!(matches!(
+            "GQ\tndls".parse::<Genotypes>(),
+            Err(ParseError::InvalidGenotype(_))
+        ));
 
         Ok(())
     }